Adjust tests after refactoring
