+++ tests/test_get_config.py
@@ -4,7 +4,8 @@ from textwrap import dedent
 import pytest
 from click.testing import CliRunner
 
-from dso.get_config import _filter_nested_dict, cli, get_config
+from dso._get_config import _filter_nested_dict, get_config
+from dso.cli import get_config_cli
 
 
 @pytest.mark.parametrize(
@@ -199,6 +200,6 @@ def test_get_config_invalid_stage(dso_project):
 def test_get_config_cli(quarto_stage):
     runner = CliRunner()
     chdir(quarto_stage)
-    result = runner.invoke(cli, ["quarto_stage"])
+    result = runner.invoke(get_config_cli, ["quarto_stage"])
     assert result.exit_code == 0
     assert "quarto:" in result.output
