+++ tests/test_lint.py
@@ -5,7 +5,8 @@ from textwrap import dedent
 import pytest
 from click.testing import CliRunner
 
-from dso.lint import DSO001, DSOLinter, LintError, QuartoRule, Rule, cli
+from dso._lint import DSO001, DSOLinter, LintError, QuartoRule, Rule
+from dso.cli import lint_cli
 
 
 @pytest.mark.parametrize(
@@ -203,6 +204,6 @@ def test_lint(quarto_stage, skip, expect_warn, expect_error):
 def test_lint_cli(dso_project, paths):
     runner = CliRunner()
     chdir(dso_project)  # proj root
-    result = runner.invoke(cli, paths)
+    result = runner.invoke(lint_cli, paths)
     print(result.output)
     assert result.exit_code == 0
